+++ src/suunto-parser.js
@@ -1,6 +1,5 @@
 'use strict';
 
 angular.module('fmd.suunto', [
-    'cb.x2js',
     'fmd.suunto.parser'
 ]);