+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.0.15
+* Removed [angular-x2js](https://github.com/cesarbarone/angular-x2js) and replaced it with [angular-x2js-wrapper](https://github.com/FrequentlyMissedDeadlines/angular-x2js-wrapper#readme)
+* Add various badges in Readme
+
 ## 1.0.9 (17/10/2020)
 * Split CI/CD to avoid publishing a new version on PR CI build.